+++ Sources/IRGen/Expression/IRBinaryExpression.swift
@@ -38,18 +38,18 @@ struct IRBinaryExpression {
 
     switch binaryExpression.opToken {
     case .plus: return IRRuntimeFunction.add(a: lhs, b: rhs)
-    case .overflowingPlus: return .functionCall(FunctionCall("add", [lhs, rhs]))
+    case .overflowingPlus: return .functionCall(FunctionCall("add", lhs, rhs))
     case .minus: return IRRuntimeFunction.sub(a: lhs, b: rhs)
-    case .overflowingMinus: return .functionCall(FunctionCall("sub", [lhs, rhs]))
+    case .overflowingMinus: return .functionCall(FunctionCall("sub", lhs, rhs))
     case .times: return IRRuntimeFunction.mul(a: lhs, b: rhs)
-    case .overflowingTimes: return .functionCall(FunctionCall("mul", [lhs, rhs]))
+    case .overflowingTimes: return .functionCall(FunctionCall("mul", lhs, rhs))
     case .divide: return IRRuntimeFunction.div(a: lhs, b: rhs)
-    case .closeAngledBracket: return .functionCall(FunctionCall("gt", [lhs, rhs]))
-    case .openAngledBracket: return .functionCall(FunctionCall("lt", [lhs, rhs]))
-    case .doubleEqual: return .functionCall(FunctionCall("eq", [lhs, rhs]))
-    case .notEqual: return .functionCall(FunctionCall("iszero", [.functionCall(FunctionCall("eq", [lhs, rhs]))]))
-    case .or: return .functionCall(FunctionCall("or", [lhs, rhs]))
-    case .and: return .functionCall(FunctionCall("and", [lhs, rhs]))
+    case .closeAngledBracket: return .functionCall(FunctionCall("gt", lhs, rhs))
+    case .openAngledBracket: return .functionCall(FunctionCall("lt", lhs, rhs))
+    case .doubleEqual: return .functionCall(FunctionCall("eq", lhs, rhs))
+    case .notEqual: return .functionCall(FunctionCall("iszero", .functionCall(FunctionCall("eq", lhs, rhs))))
+    case .or: return .functionCall(FunctionCall("or", lhs, rhs))
+    case .and: return .functionCall(FunctionCall("and", lhs, rhs))
     case .power: return IRRuntimeFunction.power(b: lhs, e: rhs)
     default: fatalError("opToken not supported")
     }
