+++ Sources/YUL/Throwing.swift
@@ -1,0 +1,10 @@
+//
+//  Throwing.swift
+//  YUL
+//
+//  Created by Aurel Bílý on 12/26/18.
+//
+
+public protocol Throwing {
+  var catchableSuccesses: [Expression] { get }
+}
