+++ Tests/UtilsTests/XCTestManifests.swift
@@ -1,0 +1,9 @@
+import XCTest
+
+#if !os(macOS)
+public func allTests() -> [XCTestCaseEntry] {
+    return [
+        testCase(UtilsTests.allTests)
+    ]
+}
+#endif
